Make sure it also works when included as an npm module
When I uploaded the package to npm, the index.js in the package was
itself placed under node_modules, which made the exclude-node-modules
feature not work at all (so 'use strict' would be applied to all files.

Now we use a substring of __dirname up to the first occurrence of the
string 'node_modules' as the base directory name.